feat: revalidation of individual post page
