+++ .eslintrc.json
@@ -2,6 +2,7 @@
   "extends": ["next/core-web-vitals","prettier"],
   "rules": {
     "react/no-unescaped-entities": "error",
-    "no-unused-vars": "warn"
+    "no-unused-vars": "warn",
+    "no-console": [ "warn", { "allow" : ["warn", "error"] }]
   }
 }
